[receiver/sqlquery]: add readme, config, and factory parts of new sql query receiver
