+++ receiver/sqlqueryreceiver/testdata/config.yaml
@@ -1,5 +1,5 @@
 receivers:
-  sql:
+  sqlquery:
     collection_interval: 10s
     driver: mydriver
     datasource: "host=localhost port=5432 user=me password=s3cr3t sslmode=disable"
@@ -15,6 +15,6 @@ service:
   pipelines:
     metrics:
       receivers:
-        - sql
+        - sqlquery
       exporters:
         - nop
