--- receiver/sqlreceiver/receiver.go
@@ -1,65 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package sqlreceiver
-
-import (
-	"context"
-	"database/sql"
-	"fmt"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/receiver/scraperhelper"
-	"go.uber.org/zap"
-)
-
-type sqlOpenerFunc func(driverName, dataSourceName string) (*sql.DB, error)
-
-type dbProviderFunc func() (*sql.DB, error)
-
-type clientProviderFunc func(*sql.DB, string, *zap.Logger) dbClient
-
-func createReceiverFunc(sqlOpenerFunc sqlOpenerFunc, clientProviderFunc clientProviderFunc) component.CreateMetricsReceiverFunc {
-	return func(
-		ctx context.Context,
-		settings component.ReceiverCreateSettings,
-		cfg config.Receiver,
-		consumer consumer.Metrics,
-	) (component.MetricsReceiver, error) {
-		sqlCfg := cfg.(*Config)
-		var opts []scraperhelper.ScraperControllerOption
-		for i, query := range sqlCfg.Queries {
-			id := config.NewComponentIDWithName("sqlreceiver", fmt.Sprintf("query-%d: %s", i, query.SQL))
-			mp := &scraper{
-				id:     id,
-				query:  query,
-				logger: settings.TelemetrySettings.Logger,
-				dbProviderFunc: func() (*sql.DB, error) {
-					return sqlOpenerFunc(sqlCfg.Driver, sqlCfg.DataSource)
-				},
-				clientProviderFunc: clientProviderFunc,
-			}
-			opt := scraperhelper.AddScraper(mp)
-			opts = append(opts, opt)
-		}
-		return scraperhelper.NewScraperControllerReceiver(
-			&sqlCfg.ScraperControllerSettings,
-			settings,
-			consumer,
-			opts...,
-		)
-	}
-}
